improve nomination UI for small viewports
issue: #326
